[#noissue] Split BatchApplicationIndexService into v1/v2 beans
Replace the readV2 flag branching inside BatchApplicationIndexServiceImpl
with two ConditionalOnProperty beans so each implementation depends only
on the DAOs it actually uses, and declare the batch service beans
explicitly instead of component scanning. The v1/v2 selection lives in
BatchApplicationIndexServiceConfig so downstream modules can import the
same wiring.
