+++ batch/src/main/java/com/navercorp/pinpoint/batch/service/BatchApplicationIndexServiceConfig.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2026 NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.navercorp.pinpoint.batch.service;
+
+import com.navercorp.pinpoint.web.dao.AgentIdDao;
+import com.navercorp.pinpoint.web.dao.ApplicationDao;
+import com.navercorp.pinpoint.web.dao.ApplicationIndexDao;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Selects the {@link BatchApplicationIndexService} implementation at the bean
+ * level with the {@code pinpoint.batch.application.index.read.v2} property.
+ */
+@Configuration
+public class BatchApplicationIndexServiceConfig {
+
+    @Bean("batchApplicationIndexService")
+    @ConditionalOnProperty(name = "pinpoint.batch.application.index.read.v2", havingValue = "false")
+    public BatchApplicationIndexService batchApplicationIndexV1Service(ApplicationIndexDao applicationIndexDao) {
+        return new BatchApplicationIndexV1ServiceImpl(applicationIndexDao);
+    }
+
+    @Bean("batchApplicationIndexService")
+    @ConditionalOnProperty(name = "pinpoint.batch.application.index.read.v2", havingValue = "true", matchIfMissing = true)
+    public BatchApplicationIndexService batchApplicationIndexV2Service(ApplicationDao applicationDao, AgentIdDao agentIdDao) {
+        return new BatchApplicationIndexV2ServiceImpl(applicationDao, agentIdDao);
+    }
+}
