+++ batch/src/main/java/com/navercorp/pinpoint/batch/service/BatchServiceConfig.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2026 NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.navercorp.pinpoint.batch.service;
+
+import com.navercorp.pinpoint.web.service.component.ActiveAgentValidator;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+@Configuration
+@Import(BatchApplicationIndexServiceConfig.class)
+public class BatchServiceConfig {
+
+    @Bean
+    public BatchAgentService batchAgentService(
+            ActiveAgentValidator activeAgentValidator,
+            BatchApplicationIndexService batchApplicationIndexService) {
+        return new BatchAgentServiceImpl(activeAgentValidator, batchApplicationIndexService);
+    }
+}
